Внесены правки
верный формат у файлов с компонентом
айдишник у сообщения генерируется автоматически
переписан useEffect
инпут стал контролируемой формой
